backend/requirements.txt
flask~=3.0.2
requests~=2.31.0
beautifulsoup4
python-dotenv~=1.0.1
pandas~=2.2.1
#flask_sqlalchemy
web3~=6.15.1
pyperclip~=1.8.2
selenium~=4.18.1